+++ docs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v1.0.0-alpha.4
 
+### Added
+
+- Added `device_connection_changed` signal to `FrayInput`
+
+- Added `Controller` node to input module.
+
 ### Changed
 
  - Completely reworked state management. The state module now includes a generic hiearchical state machine which is accessible to the user.